Make group of() private
